pyproject.toml
[tool.poetry]
name = "xray-intensity"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "xray_intensity"}]

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^9.4.0"
matplotlib = "^3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
